mercadocentral/requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.2.1
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
webencodings==0.5.1
